Move contract-escrow to services/onchain — it is the escrow contract's client
The FinP2PEscrow contract ships with the orchestrator set; its client is
on-chain domain, not custody. CustodyTokenService keeps only the optional
injection point.
