+++ src/services/custody/index.ts
@@ -1,4 +1,3 @@
 export * from './custody-provider'
 export * from './custody-registry'
 export * from './token-service'
-export * from './contract-escrow'
