+++ src/services/onchain/index.ts
@@ -5,3 +5,4 @@ export * from './plan-translator'
 export * from './plan-approval'
 export * from './proof-sync'
 export * from './services'
+export * from './contract-escrow'
